# lib/stocks-api/openapi.yaml
openapi: 3.0.3
info:
  title: Stock P/E Calculator API
  description: |
    REST API for retrieving stock quotes, historical price data, and calculating historical P/E ratios.
    
    This API uses Yahoo Finance data to provide:
    - Current stock quotes with financial metrics
    - Historical price data with customizable date ranges and intervals
    - Historical P/E ratio calculations with trailing twelve-month (TTM) EPS
    
    ## Rate Limiting
    This API relies on Yahoo Finance data. Please be respectful of rate limits and avoid excessive requests.
    
    ## Data Intervals
    - `1d` - Daily data (best for short-term analysis, < 1 year)
    - `1wk` - Weekly data (good for medium-term analysis, 1-3 years)
    - `1mo` - Monthly data (best for long-term analysis, > 3 years)
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /filings/download:
    get:
      summary: Download Filing
      description: Download the content of a specific filing
      tags:
        - SEC Filings
      parameters:
        - name: url
          in: query
          required: true
          description: URL of the filing document
          schema:
            type: string
      responses:
        '200':
          description: Filing content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            text/html:
              schema:
                type: string
            application/xml:
              schema:
                type: string
        '400':
          description: Missing or invalid URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Invalid stock symbol
        symbol:
          type: string
          example: INVALID
...